requirements.txt
streamlit>=1.35.0
psutil>=5.9.8
pandas>=2.2.0
plotly>=5.22.0
scikit-learn>=1.4.2
anthropic>=0.28.0
python-dotenv>=1.0.1
numpy>=1.26.4